libs: codecparsers: H266 GstH266RefPicListStruct's abs_delta_poc_st should be 16 bits
Its value range is 0~(2^15 − 1) according to the spec.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/5710>

+++ subprojects/gst-plugins-bad/gst-libs/gst/codecparsers/gsth266parser.h
@@ -1116,7 +1116,7 @@ struct _GstH266RefPicListStruct {
   guint8 ltrp_in_header_flag;
   guint8 inter_layer_ref_pic_flag[GST_H266_MAX_REF_ENTRIES];
   guint8 st_ref_pic_flag[GST_H266_MAX_REF_ENTRIES];
-  guint8 abs_delta_poc_st[GST_H266_MAX_REF_ENTRIES];
+  guint16 abs_delta_poc_st[GST_H266_MAX_REF_ENTRIES];
   guint8 strp_entry_sign_flag[GST_H266_MAX_REF_ENTRIES];
   guint8 rpls_poc_lsb_lt[GST_H266_MAX_REF_ENTRIES];
   guint num_short_term_pic;
